close lid: acpid enable
Maybe it's just me (and even I should know better), but I was caught out without acpid enabled and ran down my battery pretty badly when i thought it was suspended. The backlight goes off in my configuration on lid close anyway (after a delay) and I had taken that to be a suspend, but it is just the backlight.